+++ CHANGES.md
@@ -1,6 +1,13 @@
 CHANGES
 ========
 
+1.2 (2022-04-07)
+-----------------
+
+- Add __getitem__ accessor for getting the scheduling info for an event
+- Rename "postpone" into "reschedule" and always resched the same time
+- Add more error handling
+
 1.1 (2022-04-06)
 -----------------